feat: Add Google authentication functionality to auth routes
This commit adds Google authentication functionality to the auth routes. It includes the addition of a new route "/google" for handling Google authentication requests. The code makes use of the "lib/google" module and the "axios" package to authenticate with Google and retrieve user information. The user's Google credentials are stored in the database and a JWT token is generated for authentication. This feature allows users to sign in using their Google accounts, enhancing the authentication options available to users.