feat: Add optional return of Google tokens in findUserByUuid
This commit modifies the findUserByUuid function in the user schema to include an optional parameter "returnGoogleTokens". When set to true, the function will return the user's Google access token, refresh token, and token expiration date along with other user information. This change is part of the larger feature implementation for Google authentication, which includes updates to the auth routes, lib/google module, and user schema.