chore: Add migration for funny thunderbird column in user table
This commit adds a new migration file (0013_funny_thunderbird.sql) that adds a new column called "preferred_currency_id" to the "user" table. The column has a default value of '6ca4676c-2612-4a1e-a0cd-661b03b910d5' and is set to be not null. Additionally, a foreign key constraint is added to ensure referential integrity with the "currency" table.

The purpose of this migration is to introduce the "preferred_currency_id" column to store the preferred currency for each user. This will allow for more personalized user experiences and better currency management within the application.