feat: Add preferred currency field to user table
This commit adds a new column called "preferred_currency_id" to the "user" table in the `user.ts` file. The column is used to store the preferred currency for each user and is set to be not null with a default value of '6ca4676c-2612-4a1e-a0cd-661b03b910d5'. Additionally, a foreign key constraint is added to ensure referential integrity with the "currency" table.

These changes improve the user experience by allowing for personalized currency settings and better currency management within the application.